move showRemoteReverse to utils
