+++ app/src/main/kotlin/io/treehouses/remote/ui/network/NetworkViewModel.kt
@@ -92,24 +92,11 @@ class NetworkViewModel(application: Application) : FragmentViewModel(application
                 showNetworkProgress.value = false
             }
             RESULTS.REVERSE_LOOKUP ->{
-                showRemoteReverse(output)
+                Utils.showRemoteReverse(output, reverseText)
             }
             else -> logE("NewNetworkFragment: Result not Found")
         }
 
-    }
-
-    fun showRemoteReverse(output: String){
-        val reverseData = Gson().fromJson(output, ReverseData::class.java)
-        val ip = "ip: " + reverseData.ip
-        val postal = "postal: " + reverseData.postal
-        val city = "city: " + reverseData.city
-        val country = "country: " + reverseData.country
-        val org = "org: " + reverseData.org
-        val timezone = "timezone: " + reverseData.timezone
-        reverseText.value = ip + "\n" + org  + "\n" + country + "\n" + city + "\n" + postal + "\n" + timezone
-
-//        reverseText.value = output
     }
 
      fun getNetworkMode() {
